--- app/src/main/java/com/w36495/randomrithm/data/entity/Aliase.kt
@@ -1,10 +1,0 @@
-package com.w36495.randomrithm.data.entity
-
-import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class Aliase(
-    @Json(name = "alias")
-    val alias: String
-)
