--- app/src/main/java/com/w36495/randomrithm/data/entity/DisplayName.kt
@@ -1,14 +1,0 @@
-package com.w36495.randomrithm.data.entity
-
-import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class DisplayName(
-    @Json(name = "language")
-    val language: String,
-    @Json(name = "name")
-    val name: String,
-    @Json(name = "short")
-    val short: String
-)
