--- app/src/main/java/com/w36495/randomrithm/data/entity/Metadata.kt
@@ -1,6 +1,0 @@
-package com.w36495.randomrithm.data.entity
-
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-class Metadata
