+++ app/src/main/java/com/w36495/randomrithm/data/entity/ProblemDTO.kt
@@ -4,6 +4,14 @@ import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
+data class ProblemDTO(
+    @Json(name = "count")
+    val count: Int,
+    @Json(name = "items")
+    val items: List<ProblemItem>
+)
+
+@JsonClass(generateAdapter = true)
 data class ProblemItem(
     @Json(name = "acceptedUserCount")
     val acceptedUserCount: Int,
@@ -35,4 +43,40 @@ data class ProblemItem(
     val titles: List<Title>,
     @Json(name = "votedUserCount")
     val votedUserCount: Int
+)
+@JsonClass(generateAdapter = true)
+data class Title(
+    @Json(name = "isOriginal")
+    val isOriginal: Boolean,
+    @Json(name = "language")
+    val language: String,
+    @Json(name = "languageDisplayName")
+    val languageDisplayName: String,
+    @Json(name = "title")
+    val title: String
+)
+
+@JsonClass(generateAdapter = true)
+data class Tag(
+    @Json(name = "aliases")
+    val aliases: List<Aliase>,
+    @Json(name = "bojTagId")
+    val bojTagId: Int,
+    @Json(name = "displayNames")
+    val displayNames: List<DisplayName>,
+    @Json(name = "isMeta")
+    val isMeta: Boolean,
+    @Json(name = "key")
+    val key: String,
+    @Json(name = "problemCount")
+    val problemCount: Int
+)
+
+@JsonClass(generateAdapter = true)
+class Metadata
+
+@JsonClass(generateAdapter = true)
+data class Aliase(
+    @Json(name = "alias")
+    val alias: String
 )
