+++ app/src/main/java/com/w36495/randomrithm/data/remote/endpoints/AlgorithmAPI.kt
@@ -1,15 +1,15 @@
 package com.w36495.randomrithm.data.remote.endpoints
 
-import com.w36495.randomrithm.data.entity.Algorithm
-import com.w36495.randomrithm.data.entity.Problem
+import com.w36495.randomrithm.data.entity.AlgorithmDTO
+import com.w36495.randomrithm.data.entity.ProblemDTO
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Query
 
 interface AlgorithmAPI {
     @GET("tag/list")
-    fun getCountOfAlgorithm(): Call<Algorithm>
+    fun getCountOfAlgorithm(): Call<AlgorithmDTO>
 
     @GET("search/problem")
-    fun getProblemList(@Query("query") tag: String): Call<Problem>
+    fun getProblemList(@Query("query") tag: String): Call<ProblemDTO>
 }
